+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] ##
 - Added event `531`. Undocumented in the RFC. (You cannot send CTCPs
   to this user whilst they have the +T `u_noctcp` mode set.)
+- Added new [Libera Chat](https://libera.chat/) servers
 - Fixed unchecked return values
 - Made improvements to the following files:
   - `icb.c`